Move validate_or_initialize_database_version to separate file
